Moved all existing routes to route group `(site)`
Moved all existing routes to route group `(site)`

See https://kit.svelte.dev/docs/advanced-routing#advanced-layouts
