+++ src/routes/(site)/[link=link].svg/+server.ts
