+++ src/routes/(site)/blog/[id].txt/+server.ts
