+++ src/routes/(site)/blog/[id]/+page.ts
