+++ src/routes/(site)/blog/[query].json/+server.ts
