+++ src/routes/(site)/events/+page.svelte
