+++ src/routes/(site)/events/[query].json/+server.ts
