+++ src/routes/(site)/events/empty-container.svelte
