+++ src/routes/(site)/events/event.svelte
