+++ src/routes/(site)/hero.svelte
