+++ src/routes/(site)/privacy/+page.svelte
