+++ src/routes/(site)/quiz/more-info.svelte
