+++ src/routes/(site)/quiz/progress-bar.svelte
