+++ src/routes/(site)/teams/path-section.svelte
